Closes #51641. Migrating HiddenProviderTest
